Use forge-diff-visit-file in hunk section
Originally when pressing enter in a hunk section, it opens the file at
the right position in the buffer.
With a magit-diff buffer which contains diff posts, this feature is
broken, the position in the file buffer is not correct.

This patch handle this case when the magit-diff buffer contains diff
posts.

Signed-off-by: Julien Masson <massonju.eseo@gmail.com>